+++ centometer-client/app/layout.tsx
@@ -17,7 +17,7 @@ const fontBody = Manrope({
 });
 
 export const metadata: Metadata = {
-  title: "Create Next App",
+  title: "Centometer",
   description: "Generated by create next app",
 };
