+++ centometer-client/components/landing-nav.tsx
@@ -7,68 +7,18 @@ export default function LandingNavbar(){
         <header className="bg-background sticky top-0 z-40 w-full border-b border-border">
           <div className="container mx-auto flex h-16 max-w-7xl items-center justify-between px-4 sm:px-6 lg:px-8">
             <Link href="#" className="flex items-center gap-2" prefetch={false}>
-              <MountainIcon className="h-6 w-6" />
-              <span className="font-bold">Acme Inc</span>
+              <BarChartIcon className="h-6 w-6" />
+              <span className="font-bold">Centometer</span>
             </Link>
-            <nav className="hidden items-center gap-6 md:flex">
-              <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                Home
-              </Link>
-              <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                About
-              </Link>
-              <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                Services
-              </Link>
-              <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                Contact
-              </Link>
-            </nav>
             <div className="flex items-center gap-4">
               <Link
                 href="#"
-                className="hidden rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground transition-colors hover:bg-primary/90 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring md:inline-flex"
+                className="rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground transition-colors hover:bg-primary/90 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
                 prefetch={false}
               >
                 Get Started
               </Link>
-              <Sheet>
-                <SheetTrigger asChild>
-                  <Button variant="outline" size="icon" className="md:hidden">
-                    <MenuIcon className="h-6 w-6" />
-                    <span className="sr-only">Toggle navigation</span>
-                  </Button>
-                </SheetTrigger>
-                <SheetContent side="right" className="w-full max-w-xs">
-                  <div className="flex flex-col gap-6 p-6">
-                    <Link href="#" className="flex items-center gap-2" prefetch={false}>
-                      <MountainIcon className="h-6 w-6" />
-                      <span className="font-bold">Acme Inc</span>
-                    </Link>
-                    <nav className="grid gap-4">
-                      <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                        Home
-                      </Link>
-                      <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                        About
-                      </Link>
-                      <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                        Services
-                      </Link>
-                      <Link href="#" className="text-sm font-medium hover:text-primary" prefetch={false}>
-                        Contact
-                      </Link>
-                    </nav>
-                    <Link
-                      href="#"
-                      className="inline-flex w-full items-center justify-center rounded-md bg-primary px-4 py-2 text-sm font-medium text-primary-foreground transition-colors hover:bg-primary/90 focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring"
-                      prefetch={false}
-                    >
-                      Get Started
-                    </Link>
-                  </div>
-                </SheetContent>
-              </Sheet>
+             
             </div>
           </div>
         </header>
@@ -116,6 +66,28 @@ function MenuIcon(props:any) {
     )
   }
   
+  function BarChartIcon(props:any) {
+    return (
+      <svg
+        {...props}
+        xmlns="http://www.w3.org/2000/svg"
+        width="24"
+        height="24"
+        viewBox="0 0 24 24"
+        fill="none"
+        stroke="currentColor"
+        strokeWidth="2"
+        strokeLinecap="round"
+        strokeLinejoin="round"
+      >
+        <line x1="12" x2="12" y1="20" y2="10" />
+        <line x1="18" x2="18" y1="20" y2="4" />
+        <line x1="6" x2="6" y1="20" y2="16" />
+      </svg>
+    )
+  }
+
+
   
   function XIcon(props:any) {
     return (
